% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSCE.R
\name{cleanSCE}
\alias{cleanSCE}
\title{Cleans a SCE to remove all cells with zero libraries}
\usage{
cleanSCE(sce)
}
\arguments{
\item{sce}{the sce to clean}
}
\description{
This function takes an SCE, removes all cells which have no library (no non-zero entries in its column), and returns the SCE. Used to clean the SCE for later processing
}
\keyword{clean}
\keyword{sce}
