% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function_add_one.R
\name{test_function_add_one}
\alias{test_function_add_one}
\title{A Test Function}
\usage{
test_function_add_one(number = 7)
}
\arguments{
\item{number}{The number to add 1 to. Defaults to 7}
}
\description{
This function adds one to a number.
}
\examples{
test_function_add_one()
}
\keyword{test}
